Fix issue with local package name
SPM uses the name of the directory where Package.swift is located as
the name of a local package. This change makes the package manifest
independent of  the name of the directory to which the repository is
cloned. This should fix an issue where package dependencies can't be
resolved if the repository is cloned to a directory with a name
different from "swift-shell".